+++ Tool/kvsconv/main.cpp
@@ -1,26 +1,25 @@
 /*****************************************************************************/
 /**
  *  @file   main.cpp
- */
-/*----------------------------------------------------------------------------
- *
- *  Copyright (c) Visualization Laboratory, Kyoto University.
- *  All rights reserved.
- *  See http://www.viz.media.kyoto-u.ac.jp/kvs/copyright/ for details.
- *
- *  $Id: main.cpp 631 2010-10-10 02:15:35Z naohisa.sakamoto $
+ *  @author Naohisa Sakamoto
  */
 /*****************************************************************************/
 #include <kvs/MemoryDebugger>
+#include <kvs/Message>
 #include "main.h"
 #include "Argument.h"
-#include "fld2kvsml.h"
-#include "ucd2kvsml.h"
-#include "img2img.h"
-#include "tet2tet.h"
-#include <kvs/Message>
+#include "FldConv.h"
+#include "UcdConv.h"
+#include "TetConv.h"
+#include "ImgConv.h"
 
 KVS_MEMORY_DEBUGGER;
+
+#define KVSCONV_HELP( command )                                         \
+    if ( c == command::CommandName ) return command::Argument( m_argc, m_argv ).parse()
+
+#define KVSCONV_EXEC( command )                                         \
+    if ( arg.hasOption( command::CommandName ) ) return command::Main( m_argc, m_argv ).exec()
 
 
 namespace kvsconv
@@ -40,22 +39,20 @@ bool Main::exec()
     // Output help message for each converter.
     if ( arg.hasOption("help") )
     {
-        const std::string c = arg.optionValue<std::string>("help");
-        if ( c == "fld2kvsml" ) return fld2kvsml::Argument( m_argc, m_argv ).parse();
-        if ( c == "ucd2kvsml" ) return ucd2kvsml::Argument( m_argc, m_argv ).parse();
-        if ( c == "img2img" ) return img2img::Argument( m_argc, m_argv ).parse();
-        if ( c == "tet2tet" ) return tet2tet::Argument( m_argc, m_argv ).parse();
-
-        kvsMessageError( "Unknown converter %s.", c.c_str() );
+        const auto c = arg.optionValue<std::string>("help");
+        KVSCONV_HELP( FldConv );
+        KVSCONV_HELP( UcdConv );
+        KVSCONV_HELP( TetConv );
+        KVSCONV_HELP( ImgConv );
+        kvsMessageError() << "Unknown converter '" << c << "'." << std::endl;;
         return false;
     }
 
     // Convert.
-    if ( arg.hasOption("fld2kvsml") ) return fld2kvsml::Main( m_argc, m_argv ).exec();
-    if ( arg.hasOption("ucd2kvsml") ) return ucd2kvsml::Main( m_argc, m_argv ).exec();
-    if ( arg.hasOption("img2img") ) return img2img::Main( m_argc, m_argv ).exec();
-    if ( arg.hasOption("tet2tet") ) return tet2tet::Main( m_argc, m_argv ).exec();
-
+    KVSCONV_EXEC( FldConv );
+    KVSCONV_EXEC( UcdConv );
+    KVSCONV_EXEC( TetConv );
+    KVSCONV_EXEC( ImgConv );
     return false;
 }